pyproject.toml
[tool.poetry]
name = "stockmarket"
version = "0.1.0"
description = "\"Stock Market Analysis\""
authors = ["Sebastian Carrillo Mejia <sebastian.carrillo1@udea.edu.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.37.1"
plotly = "^5.23.0"
pandas = "^2.2.2"
yfinance = "^0.2.42"
matplotlib = "^3.9.2"
seaborn = "^0.13.2"
python-dateutil = "^2.9.0.post0"
tables = "^3.10.1"
scikit-learn = "^1.5.1"
mlflow = "^2.16.0"
numpy = "1.26.4"
tensorflow = "^2.17.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
